改动
来电宝接口 微信服务号关注自动回复图文  修理厂列表上传excel 微信推送消息模板 redis 引入

+++ CarRepairManage/Repository/Base/IBaseRepository.cs
@@ -32,7 +32,7 @@ namespace Repository
 
         long Insert(T entity);
 
-        int Update(T entity);
+        long Update(T entity);
 
         int Delete(T entity);
